fix balloon emoji for crates.io and docs.rs
The `:balloon:` emoji directive works when the markdown is rendered on
GitHub. However, the emoji 🎈 doesn't display on crates.io or docs.rs.

This change replaces the directive with the emoji. We are lucky to live
in 2023 where this will probably work in all (modern enough) editors.

Some miscellaneous typos were also fixed.